# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unworn — a marketplace for brand-new, never-worn clothing.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bilcost/search:
    get:
      operationId: bilcostSearchListings
      tags: [bilcost]
      summary: Search Bilbasen-style listings (curated catalog in v1)
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: make
          in: query
          required: false
          schema: { type: string }
        - name: model
          in: query
          required: false
          schema: { type: string }
        - name: isNew
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BilcostListingPreview"
  /bilcost/listing:
    get:
      operationId: bilcostGetListing
      tags: [bilcost]
      summary: Resolve a Bilbasen listing URL or catalog id into a Car
      parameters:
        - name: urlOrId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BilcostCar"
        "400":
          description: Could not resolve listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /bilcost/compare:
    post:
      operationId: bilcostCompareCars
      tags: [bilcost]
      summary: Compute side-by-side TCO for two cars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BilcostCompareInput"
      responses:
        "200":
          description: Comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BilcostComparison"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    BilcostFuelType:
      type: string
      enum: [petrol, diesel, hybrid, phev, electric]
    BilcostTransmission:
      type: string
      enum: [manual, automatic, unknown]
    BilcostDriverAgeBand:
      type: string
      enum: [under_25, 25_29, 30_49, 50_64, 65_plus]
    BilcostListingPreview:
      type: object
      required: [id, make, model, year, mileageKm, priceDkk, fuelType, listingUrl]
      properties:
        id: { type: string }
        make: { type: string }
        model: { type: string }
        variant:
          type: ["string", "null"]
        year: { type: integer }
        mileageKm: { type: integer }
        priceDkk: { type: integer }
        fuelType:
          $ref: "#/components/schemas/BilcostFuelType"
        imageUrl:
          type: ["string", "null"]
        listingUrl: { type: string }
    BilcostCar:
      type: object
      required:
        [
          make,
          model,
          year,
          mileageKm,
          priceDkk,
          fuelType,
          transmission,
        ]
      properties:
        make: { type: string }
        model: { type: string }
        variant:
          type: ["string", "null"]
        year: { type: integer }
        mileageKm: { type: integer }
        priceDkk: { type: integer }
        fuelType:
          $ref: "#/components/schemas/BilcostFuelType"
        consumptionKmPerL:
          type: ["number", "null"]
        consumptionKwhPer100km:
          type: ["number", "null"]
        co2Gkm:
          type: ["integer", "null"]
        kerbWeightKg:
          type: ["integer", "null"]
        transmission:
          $ref: "#/components/schemas/BilcostTransmission"
        listingUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    BilcostAssumptions:
      type: object
      required:
        [
          yearsHorizon,
          annualKm,
          downPaymentDkk,
          loanInterestPct,
          loanTermYears,
          fuelPriceDkkPerLiter,
          electricityPriceDkkPerKwh,
          driverAgeBand,
          postalCode,
        ]
      properties:
        yearsHorizon: { type: integer, minimum: 1, maximum: 15 }
        annualKm: { type: integer, minimum: 1000, maximum: 80000 }
        downPaymentDkk: { type: integer, minimum: 0 }
        loanInterestPct: { type: number, minimum: 0, maximum: 30 }
        loanTermYears: { type: integer, minimum: 0, maximum: 15 }
        fuelPriceDkkPerLiter: { type: number, minimum: 0, maximum: 50 }
        electricityPriceDkkPerKwh: { type: number, minimum: 0, maximum: 20 }
        insuranceAnnualDkkOverride:
          type: ["integer", "null"]
        driverAgeBand:
          $ref: "#/components/schemas/BilcostDriverAgeBand"
        postalCode: { type: string }
    BilcostCostComponent:
      type: object
      required: [amountDkk, perYearDkk, notes]
      properties:
        amountDkk: { type: integer }
        perYearDkk:
          type: array
          items: { type: integer }
        notes: { type: string }
    BilcostBreakdown:
      type: object
      required:
        [
          depreciation,
          financingInterest,
          insurance,
          maintenance,
          fuel,
          ownershipTax,
          tiresAndWear,
          inspection,
        ]
      properties:
        depreciation: { $ref: "#/components/schemas/BilcostCostComponent" }
        financingInterest: { $ref: "#/components/schemas/BilcostCostComponent" }
        insurance: { $ref: "#/components/schemas/BilcostCostComponent" }
        maintenance: { $ref: "#/components/schemas/BilcostCostComponent" }
        fuel: { $ref: "#/components/schemas/BilcostCostComponent" }
        ownershipTax: { $ref: "#/components/schemas/BilcostCostComponent" }
        tiresAndWear: { $ref: "#/components/schemas/BilcostCostComponent" }
        inspection: { $ref: "#/components/schemas/BilcostCostComponent" }
    BilcostTcoResult:
      type: object
      required:
        [
          car,
          assumptions,
          breakdown,
          totalDkk,
          totalPerYearDkk,
          totalPerKmDkk,
          residualValueDkk,
          yearlyCumulativeDkk,
        ]
      properties:
        car: { $ref: "#/components/schemas/BilcostCar" }
        assumptions: { $ref: "#/components/schemas/BilcostAssumptions" }
        breakdown: { $ref: "#/components/schemas/BilcostBreakdown" }
        totalDkk: { type: integer }
        totalPerYearDkk: { type: integer }
        totalPerKmDkk: { type: number }
        residualValueDkk: { type: integer }
        yearlyCumulativeDkk:
          type: array
          items: { type: integer }
    BilcostComparison:
      type: object
      required: [carA, carB, deltaDkk]
      properties:
        carA: { $ref: "#/components/schemas/BilcostTcoResult" }
        carB: { $ref: "#/components/schemas/BilcostTcoResult" }
        deltaDkk: { type: integer }
    BilcostCompareInput:
      type: object
      required: [carA, carB, assumptions]
      properties:
        carA: { $ref: "#/components/schemas/BilcostCar" }
        carB: { $ref: "#/components/schemas/BilcostCar" }
        assumptions: { $ref: "#/components/schemas/BilcostAssumptions" }