# lib/api-zod/src/generated/types/bilcostAssumptions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unworn — a marketplace for brand-new, never-worn clothing.
 * OpenAPI spec version: 0.1.0
 */
import type { BilcostDriverAgeBand } from "./bilcostDriverAgeBand";

export interface BilcostAssumptions {
  /**
   * @minimum 1
   * @maximum 15
   */
  yearsHorizon: number;
  /**
   * @minimum 1000
   * @maximum 80000
   */
  annualKm: number;
  /** @minimum 0 */
  downPaymentDkk: number;
  /**
   * @minimum 0
   * @maximum 30
   */
  loanInterestPct: number;
  /**
   * @minimum 0
   * @maximum 15
   */
  loanTermYears: number;
  /**
   * @minimum 0
   * @maximum 50
   */
  fuelPriceDkkPerLiter: number;
  /**
   * @minimum 0
   * @maximum 20
   */
  electricityPriceDkkPerKwh: number;
  /** @nullable */
  insuranceAnnualDkkOverride?: number | null;
  driverAgeBand: BilcostDriverAgeBand;
  postalCode: string;
}