# lib/api-zod/src/generated/types/bilcostComparison.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unworn — a marketplace for brand-new, never-worn clothing.
 * OpenAPI spec version: 0.1.0
 */
import type { BilcostTcoResult } from "./bilcostTcoResult";

export interface BilcostComparison {
  carA: BilcostTcoResult;
  carB: BilcostTcoResult;
  deltaDkk: number;
}